% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichDotplot.R
\name{enrichDotplot}
\alias{enrichDotplot}
\title{Enrichment Dot-Plot Server}
\usage{
enrichDotplot(input, output, session)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}
}
\description{
Enrichment Dot-Plot Server
}
