% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichDotplot.R
\name{enrichDotplotInput}
\alias{enrichDotplotInput}
\title{Enrichment Dot-Plot UI.}
\usage{
enrichDotplotInput(id)
}
\arguments{
\item{id}{Input id for use in server functions.}
}
\description{
Enrichment Dot-Plot UI.
}
