% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichDotplot.R
\name{enrichDotplotServer}
\alias{enrichDotplotServer}
\title{Enrichment Dot-Plot Server}
\usage{
enrichDotplotServer(id, clusters = "seurat_clusters")
}
\arguments{
\item{id}{Input id for use in server functions.}

\item{clusters}{Cluster column name.}
}
\description{
Enrichment Dot-Plot Server
}
