% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichDotplot.R
\name{enrichDotplotUI}
\alias{enrichDotplotUI}
\title{Enrichment Dot-Plot UI.}
\usage{
enrichDotplotUI(id)
}
\arguments{
\item{id}{Input id for use in server functions.}
}
\description{
Enrichment Dot-Plot UI.
}
