% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichTable.R
\name{enrichTableServer}
\alias{enrichTableServer}
\title{Enrichment Table Server}
\usage{
enrichTableServer(id, clusters = "seurat_clusters")
}
\arguments{
\item{id}{Input id for use in server functions.}

\item{clusters}{Cluster column name.}
}
\description{
Enrichment Table Server
}
