% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichTable.R
\name{enrichTableUI}
\alias{enrichTableUI}
\title{Term Enrichment Table UI}
\usage{
enrichTableUI(id)
}
\arguments{
\item{id}{Input id for use in server functions.}
}
\description{
Term Enrichment Table UI
}
