% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDimPlot.R
\name{expDimPlot}
\alias{expDimPlot}
\title{Expression Dim Plot Server}
\usage{
expDimPlot(
  input,
  output,
  session,
  ident = "orig.ident",
  clusters = "seurat_clusters"
)
}
\description{
Expression Dim Plot Server
}
