% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDimPlot.R
\name{expDimPlotInput}
\alias{expDimPlotInput}
\title{Expression Dim Plot UI}
\usage{
expDimPlotInput(id, ident = "orig.ident", clusters = "seurat_clusters")
}
\arguments{
\item{id}{Input id for use in server functions.}

\item{ident}{Sample identity column name.}

\item{clusters}{Cluster column name.}
}
\description{
Expression Dim Plot UI
}
