% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputOptions.R
\name{experimentInput}
\alias{experimentInput}
\title{Experiment Selector}
\usage{
experimentInput(id, experiments)
}
\arguments{
\item{id}{Input id for use in server functions.}

\item{experiments}{Experiments to choose from.}
}
\description{
Experiment Selector
}
