% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expPlot.R
\name{expPlot}
\alias{expPlot}
\title{Expression Plot Server.}
\usage{
expPlot(
  input,
  output,
  session,
  ident = "orig.ident",
  clusters = "seurat_clusters"
)
}
\arguments{
\item{ident}{Sample identity column name.}

\item{clusters}{Cluster column name.}
}
\description{
Expression Plot Server.
}
