% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expPlot.R
\name{expPlotServer}
\alias{expPlotServer}
\title{Expression Plot Server.}
\usage{
expPlotServer(id, ident = "orig.ident", clusters = "seurat_clusters")
}
\arguments{
\item{id}{Input id for use in server functions.}

\item{ident}{Sample identity column name.}

\item{clusters}{Cluster column name.}
}
\description{
Expression Plot Server.
}
