% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expPlot.R
\name{expPlotUI}
\alias{expPlotUI}
\title{Expression Plot UI.}
\usage{
expPlotUI(id, ident = "orig.ident", clusters = "seurat_clusters")
}
\arguments{
\item{id}{Input id for use in server functions.}

\item{ident}{Sample identity column name.}

\item{clusters}{Cluster column name.}
}
\description{
Expression Plot UI.
}
