% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expTable.R
\name{expTable}
\alias{expTable}
\title{Expression Table Server}
\usage{
expTable(
  input,
  output,
  session,
  ident = "orig.ident",
  clusters = "seurat_clusters"
)
}
\description{
Expression Table Server
}
