% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expTable.R
\name{expTableServer}
\alias{expTableServer}
\title{Expression Table Server}
\usage{
expTableServer(id, ident = "orig.ident", clusters = "seurat_clusters")
}
\arguments{
\item{id}{Input id for use in server functions.}

\item{ident}{Sample identity column name.}

\item{clusters}{Cluster column name.}
}
\description{
Expression Table Server
}
