% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerDotplot.R
\name{markerDotplotInput}
\alias{markerDotplotInput}
\title{Marker Dot-Plot UI.}
\usage{
markerDotplotInput(id, ident = "orig.ident", clusters = "seurat_clusters")
}
\arguments{
\item{id}{Input id for use in server functions.}

\item{ident}{Sample identity column name.}

\item{clusters}{Cluster column name.}
}
\description{
Marker Dot-Plot UI.
}
