% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerTable.R
\name{markerTableServer}
\alias{markerTableServer}
\title{Marker Table Server}
\usage{
markerTableServer(id, clusters = "seurat_clusters")
}
\arguments{
\item{id}{Input id for use in server functions.}

\item{clusters}{Cluster column name.}
}
\description{
Marker Table Server
}
