% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadataPlot}
\alias{metadataPlot}
\title{Meta_data Plot Input Server}
\usage{
metadataPlot(input, output, session)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}
}
\description{
Meta_data Plot Input Server
}
