% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataPlot.R
\name{metadataPlotUI}
\alias{metadataPlotUI}
\title{Meta-Data Plot Input UI}
\usage{
metadataPlotUI(id, ident = "orig.ident", clusters = "seurat_clusters")
}
\arguments{
\item{id}{Input id for use in server functions.}

\item{ident}{Sample identity column name.}

\item{clusters}{Cluster column name.}

\item{nCount}{SCT UMI counts column name.}

\item{nFeature}{SCT feature counts column name.}

\item{percentMT}{Cell mitochondrial percentage column name.}
}
\description{
Meta-Data Plot Input UI
}
