% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataTable.R
\name{metadataTable}
\alias{metadataTable}
\title{Meta-Data Table Output Server}
\usage{
metadataTable(
  input,
  output,
  session,
  ident = "orig.ident",
  clusters = "seurat_clusters",
  nCount = "nCount_SCT",
  nFeature = "nFeature_SCT",
  percentMT = "percent.mt"
)
}
\description{
Meta-Data Table Output Server
}
