% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataTable.R
\name{metadataTableInput}
\alias{metadataTableInput}
\title{Meta-Data Table Input}
\usage{
metadataTableInput(
  id,
  ident = "orig.ident",
  clusters = "seurat_clusters",
  nCount = "nCount_SCT",
  nFeature = "nFeature_SCT",
  percentMT = "percent.mt"
)
}
\description{
Meta-Data Table Input
}
