% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataTable.R
\name{metadataTableServer}
\alias{metadataTableServer}
\title{Meta-Data Table Output Server}
\usage{
metadataTableServer(id, ident = "orig.ident", clusters = "seurat_clusters")
}
\arguments{
\item{id}{Input id for use in server functions.}

\item{ident}{Sample identity column name.}

\item{clusters}{Cluster column name.}
}
\description{
Meta-Data Table Output Server
}
