% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataTable.R
\name{metadataTableUI}
\alias{metadataTableUI}
\title{Meta-Data Table Input}
\usage{
metadataTableUI(
  id,
  ident = "orig.ident",
  clusters = "seurat_clusters",
  nCount = "nCount_SCT",
  nFeature = "nFeature_SCT",
  percentMT = "percent.mt"
)
}
\arguments{
\item{id}{Input id for use in server functions.}

\item{ident}{Sample identity column name.}

\item{clusters}{Cluster column name.}

\item{nCount}{SCT UMI counts column name.}

\item{nFeature}{SCT feature counts column name.}

\item{percentMT}{Cell mitochondrial percentage column name.}
}
\description{
Meta-Data Table Input
}
