% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputOptions.R
\name{sampleInput}
\alias{sampleInput}
\title{Sample Picker}
\usage{
sampleInput(id, samples)
}
\arguments{
\item{id}{Input id for use in server functions.}

\item{samples}{available for selection.}
}
\description{
Sample Picker
}
