% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDimPlot.R
\name{expDimPlot}
\alias{expDimPlot}
\title{Expression Dim Plot Server}
\usage{
expDimPlot(
  input,
  output,
  session,
  ident = "orig.ident",
  clusters = "seurat_clusters"
)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{ident}{Sample identity column name.}

\item{clusters}{Cluster column name.}
}
\description{
Expression Dim Plot Server
}
