% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerDotplot.R
\name{markerDotplot}
\alias{markerDotplot}
\title{Marker Dot-Plot Server.}
\usage{
markerDotplot(
  input,
  output,
  session,
  ident = "orig.ident",
  clusters = "seurat_clusters"
)
}
\description{
Marker Dot-Plot Server.
}
