% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadataPlotInput}
\alias{metadataPlotInput}
\title{Meta-Data Plot Input UI}
\usage{
metadataPlotInput(id)
}
\arguments{
\item{id}{Input id for use in server functions.}
}
\description{
Meta-Data Plot Input UI
}
