% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataTable.R
\name{metadataTable}
\alias{metadataTable}
\title{Meta-Data Table Output Server}
\usage{
metadataTable(
  input,
  output,
  session,
  ident = "orig.ident",
  clusters = "seurat_clusters",
  nCount = "nCount_SCT",
  nFeature = "nFeature_SCT",
  percentMT = "percent.mt"
)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{ident}{Sample identity column name.}

\item{clusters}{Cluster column name.}

\item{nCount}{SCT UMI counts column name.}

\item{nFeature}{SCT feature counts column name.}

\item{percentMT}{Cell mitochondrial percentage column name.}
}
\description{
Meta-Data Table Output Server
}
